Fixed the communication with MQTT
Unity successfully sends and controls the physical house and vice versa. There is a bug with the living room switch where it turns off after a time out since the sensor in real life does not detect anything. Front of the house sensor in the physical house stays on, issue with the physcial component